Move indent ranges to ITokenizedModel
